requirements.txt
requests
flask
openai
python-dotenv
markdown